app/requirements.txt
smbus
pimoroni-bme280
picamera
pillow
Jinja2
azure-storage-blob
psutil
retry

